add empty firehose loss/signpost/nonactivity/trace files
