+++ extension/encoding/macosunifiedloggingencodingextension/firehose_loss.go
@@ -1,0 +1,7 @@
+package macosunifiedloggingencodingextension
+
+type FirehoseLoss struct{}
+
+func ParseFirehoseLoss(data []byte) (FirehoseLoss, []byte) {
+	return FirehoseLoss{}, data
+}
