+++ extension/encoding/macosunifiedloggingencodingextension/firehose_nonactivity.go
@@ -1,0 +1,10 @@
+package macosunifiedloggingencodingextension
+
+type FirehoseNonActivity struct {
+	PrivateStringsSize   uint16
+	PrivateStringsOffset uint16
+}
+
+func ParseFirehoseNonActivity(data []byte, flags uint16) (FirehoseNonActivity, []byte) {
+	return FirehoseNonActivity{}, data
+}
