+++ extension/encoding/macosunifiedloggingencodingextension/firehose_signpost.go
@@ -1,0 +1,7 @@
+package macosunifiedloggingencodingextension
+
+type FirehoseSignpost struct{}
+
+func ParseFirehoseSignpost(data []byte, flags uint16) (FirehoseSignpost, []byte) {
+	return FirehoseSignpost{}, data
+}
