+++ extension/encoding/macosunifiedloggingencodingextension/firehose_trace.go
@@ -1,0 +1,7 @@
+package macosunifiedloggingencodingextension
+
+type FirehoseTrace struct{}
+
+func ParseFirehoseTrace(data []byte) (FirehoseTrace, []byte) {
+	return FirehoseTrace{}, data
+}
